fix: put extended proxy behind feature flag
