+++ runtimes/runtimes/util/standalone/proxyUtil.test.ts
@@ -8,7 +8,7 @@ import { strict as assert } from 'node:assert'
 import sinon from 'sinon'
 import mockFs from 'mock-fs'
 import { HttpsProxyAgent } from 'hpagent'
-import { ProxyConfigManager } from './proxyUtil'
+import { ProxyConfigManager } from './experimentalProxyUtil'
 import * as certificatesReaders from './certificatesReaders'
 import { Telemetry } from '../../../server-interface'
 import forge from 'node-forge'
